Add shared test-utils for component tests
