+++ src/opening-period/page/CreateNewOpeningPeriodPage.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { act } from 'react-dom/test-utils';
 import { render, screen, fireEvent } from '@testing-library/react';
+import { getElementOrThrow } from '../../../test/test-utils';
 import {
   DatePeriod,
   UiDatePeriodConfig,
@@ -17,22 +18,6 @@ jest.mock('react-router-dom', () => ({
     push: mockHistoryPush,
   }),
 }));
-
-function getElementOrThrow(
-  container: Element | null,
-  selector: string
-): Element {
-  if (container === null) {
-    throw new Error('container was null');
-  }
-  const element = container.querySelector(selector);
-
-  if (!element) {
-    throw new Error(`Element with selector ${selector} not found`);
-  }
-
-  return element;
-}
 
 async function fillCompulsoryPeriodDescriptionFields(
   container: Element
