+++ tsconfig.eslint.json
@@ -2,6 +2,7 @@
   "extends": "./tsconfig.json",
   "include": [
     "src",
-    "cypress"
+    "cypress",
+    "test"
   ]
 }
